backend/requirements.txt
# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1

# LLM (opsiyonel)
anthropic==0.52.0

# Vector DB
qdrant-client==1.14.2

# Embedding (CPU-only torch)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers==4.1.0

# Database
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
alembic==1.15.2

# Cache
redis==5.3.0

# HTTP client
httpx==0.28.1

# Auth
bcrypt
PyJWT
email-validator

# Document processing
pymupdf
python-docx
python-multipart
fpdf2

# Utils
python-dotenv==1.1.0
tenacity==9.1.2
structlog==25.4.0

# Testing
pytest
pytest-asyncio
